+++ euexit/tests/test_forms.py
@@ -1,6 +1,7 @@
 from directory_constants import choices
 import pytest
 
+from core.constants import CONSENT_EMAIL
 from euexit import forms
 
 
@@ -41,7 +42,6 @@ def test_contact_form_set_field_attributes(form_class):
     form_one = form_class(
         field_attributes={},
         ingress_url='http://www.ingress.com',
-        disclaimer='disclaim',
     )
     form_two = form_class(
         field_attributes={
@@ -53,22 +53,18 @@ def test_contact_form_set_field_attributes(form_class):
             }
         },
         ingress_url='http://www.ingress.com',
-        disclaimer='disclaim',
     )
 
     assert form_one.fields['first_name'].label is None
     assert form_one.fields['last_name'].label is None
-    assert form_one.fields['terms_agreed'].widget.label.endswith('disclaim')
     assert form_two.fields['first_name'].label == 'Your given name'
     assert form_two.fields['last_name'].label == 'Your family name'
-    assert form_two.fields['terms_agreed'].widget.label.endswith('disclaim')
 
 
 def test_domestic_contact_form_serialize(captcha_stub):
     form = forms.DomesticContactForm(
         field_attributes={},
         ingress_url='http://www.ingress.com',
-        disclaimer='disclaim',
         data={
             'first_name': 'test',
             'last_name': 'example',
@@ -76,7 +72,7 @@ def test_domestic_contact_form_serialize(captcha_stub):
             'organisation_type': 'COMPANY',
             'company_name': 'thing',
             'comment': 'hello',
-            'terms_agreed': True,
+            'contact_consent': [CONSENT_EMAIL],
             'g-recaptcha-response': captcha_stub,
         }
     )
@@ -89,6 +85,7 @@ def test_domestic_contact_form_serialize(captcha_stub):
         'company_name': 'thing',
         'comment': 'hello',
         'ingress_url': 'http://www.ingress.com',
+        'contact_consent': [CONSENT_EMAIL],
     }
 
 
@@ -96,7 +93,6 @@ def test_international_contact_form_serialize(captcha_stub):
     form = forms.InternationalContactForm(
         field_attributes={},
         ingress_url='http://www.ingress.com',
-        disclaimer='disclaim',
         data={
             'first_name': 'test',
             'last_name': 'example',
@@ -122,4 +118,5 @@ def test_international_contact_form_serialize(captcha_stub):
         'city': 'London',
         'comment': 'hello',
         'ingress_url': 'http://www.ingress.com',
+        'terms_agreed': True,
     }