+++ marketaccess/forms.py
@@ -4,9 +4,16 @@ from django.utils.safestring import mark_safe
 
 from django.forms import Textarea, TextInput
 
+from marketaccess import widgets
+
 
 LOCATION_CHOICES = [('', 'Please select')] + choices.COUNTRIES_AND_TERRITORIES
-LOCATION_CHOICES_MAP = dict(LOCATION_CHOICES)
+LOCATION_MAP = dict(LOCATION_CHOICES)
+PROBLEM_CAUSE_CHOICES = (
+    ('brexit', 'Brexit'),
+    ('covid-19', 'Covid-19'),
+)
+PROBLEM_CAUSE_MAP = dict(PROBLEM_CAUSE_CHOICES)
 
 
 class AboutForm(forms.Form):
@@ -168,24 +175,30 @@ class ProblemDetailsForm(forms.Form):
             )
         }
     )
-    eu_exit_related = forms.ChoiceField(
-        label='Is your problem caused by or related to Brexit?',
-        widget=forms.RadioSelect(
-            use_nice_ids=True, attrs={'id': 'radio-one'}
+    problem_cause = forms.MultipleChoiceField(
+        label='Is the problem caused by or related to any of the following?',
+        widget=widgets.TickboxWithOptionsHelpText(
+            use_nice_ids=True,
+            attrs={
+                'id': 'radio-one',
+                'help_text': {
+                     'radio-one-covid-19': 'Problem related to the COVID-19 (coronavirus) pandemic.',
+                }
+            },
         ),
-        choices=(
-            ('Yes', 'Yes'),
-            ('No', 'No')
-        ),
-        error_messages={
-            'required': 'Tell us if your problem is related to Brexit'
-        }
+        choices=PROBLEM_CAUSE_CHOICES,
+        required=False,
     )
 
     def clean_location(self):
-        location = self.cleaned_data['location']
-        self.cleaned_data['location_label'] = LOCATION_CHOICES_MAP[location]
-        return location
+        value = self.cleaned_data['location']
+        self.cleaned_data['location_label'] = LOCATION_MAP[value]
+        return value
+
+    def clean_problem_cause(self):
+        value = self.cleaned_data['problem_cause']
+        self.cleaned_data['problem_cause_label'] = [PROBLEM_CAUSE_MAP[item] for item in value]
+        return value
 
 
 class SummaryForm(forms.Form):